Modified the way the config file is found by IOTool
Now, IOTool looks in the root directory using a relative path from itself.
This way, even if you only import a single driver (from another custom program,
or from the command line), it can still find the config file to check
whether to use prologix or pyvisa on this machine.