made the device port comboboxes editable from the Instrument Widget interface
added a virtual instrument to be able to fetch measurements made by instruments on other computers, it is a class which inherit from the Instrument class of one of the driver in the LabDriber package. The method measure is rewritten to send an IP request and return the value of the data measured by the instrument connected to the other computer.

The virtual instrument is like a client connecting to a server